Embed internal VersionNumber in GradleRuntimeCompatibility
